refactor: 구조 재구성
